chore: pull code to get all contracts up into hir context
